Handle OVN NB and SB databases with separate ovsdb-server procs
To help with scale issues, default devstack to spawn a separate
ovsdb-server process for the OVN NB database, listening on port 6641,
using a different db socket, and a different PID file for ovs-appctl.
Included are corresponding changes needed to start ovn-northd properly,
to ensure that the Neutron plugin connects correctly, and to configure
vagrant correctly.

Closes-Bug: 1542005

Change-Id: Iaa04cbe0dcf9661c9ee23f4f94e3db795b08c167
Signed-off-by: RYAN D. MOATS <rmoats@us.ibm.com>